Try to fix %zu compiling warning

+++ src/bq_common/utils/file_manager.cpp
@@ -332,8 +332,8 @@ namespace bq {
         int32_t error_code = bq::platform::write_file(handle.platform_handle(), data, length, total_written_count);
         if (error_code != 0) {
             file_manager_file_error_no_ = error_code;
-            FILE_MANAGER_LOG(bq::log_level::error, "write file failed, errno:%d, path:%s, need_write_count:%zu, real_write_count:%zu",
-                error_code, handle.file_path_.c_str(), length, total_written_count);
+            FILE_MANAGER_LOG(bq::log_level::error, "write file failed, errno:%" PRId32 ", path:%s, need_write_count:%" PRIu64 ", real_write_count:%" PRIu64,
+                error_code, handle.file_path_.c_str(), static_cast<uint64_t>(length), static_cast<uint64_t>(total_written_count));
             return total_written_count;
         }
         return total_written_count;
@@ -369,8 +369,8 @@ namespace bq {
             int32_t error_code = bq::platform::read_file(handle.platform_handle(), (uint8_t*)dest_data + total_read_count, need_read_count, tmp_count);
             if (error_code != 0) {
                 file_manager_file_error_no_ = error_code;
-                FILE_MANAGER_LOG(bq::log_level::error, "read file failed, errno:%d, path:%s, need_read_count:%zu, real_read_count:%zu",
-                    error_code, handle.file_path_.c_str(), length, total_read_count);
+                FILE_MANAGER_LOG(bq::log_level::error, "read file failed, errno:%d, path:%s, need_read_count:%" PRIu64 ", real_read_count:%" PRIu64,
+                    error_code, handle.file_path_.c_str(), static_cast<uint64_t>(length), static_cast<uint64_t>(total_read_count);
                 break;
             }
             assert(tmp_count <= need_read_count && "read_count <= length");
